src/requirements.txt
Flask==2.0.2
functions-framework==3.*
google-cloud-datastore==2.10.0
